+++ modules/kalman/ekf/include/ekf.h
@@ -21,6 +21,9 @@
 #include <math.h>
 #include <stdint.h>
 #include <mat.h>
+#include <string.h>
+#include <cstring>
+#include <cmath>
 
 /**
  * The ekf is a base class for Extended Kalman Filter.