+++ test_app/pytest_esp-dsp_unity_tests.py
@@ -4,9 +4,7 @@
 import pytest
 from pytest_embedded import Dut
 
-@pytest.mark.esp32
-@pytest.mark.esp32s3
-@pytest.mark.esp32p4
+@pytest.mark.parametrize('target', ['esp32', 'esp32s3', 'esp32p4'])
 def test_esp_system(dut: Dut) -> None:
     dut.run_all_single_board_cases()