+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] 
 ### Fixed
 - Update ranges in tests to pass for target linux
+- Bugfix fird f32 for esp32s3
 
 ## [1.7.0] 2025-06-15
 ### Added